refactor(config): split configuration into site and profile modules
