+++ src/types/site.ts
@@ -1,39 +1,16 @@
-import type { IconName } from '@/lib/icons'
-
 export type NavItem = {
   label: string
   href: string
   variant?: 'link' | 'button'
 }
 
-export type SocialLink = {
-  name: string
-  url: string
-  icon?: 'linkedin' | 'github'
-}
-
-export type SkillCategory<T = IconName> = {
+export type Section = {
   title: string
   description: string
-  stack: T[]
 }
 
-export type Skills = {
-  backend: SkillCategory
-  frontend: SkillCategory
-  devOps: SkillCategory
-  devWorkflow: SkillCategory
-  softwareArchitecture: SkillCategory<string>
-}
-
-export type Experience = {
-  company: string
-  role: string
-  period: string
-}
-
-export type Project = {
-  title: string
-  description: string
-  url: string
+export type Sections = {
+  skills: Section
+  experiences: Section
+  projects: Section
 }
